feat(settings): redesign Familiars tab + fold project permissions into per-familiar tabs
Reworks Settings → Familiars and relocates the project-permissions protocol
out of a standalone console and into each familiar's own studio.

- Familiars picker: replace the plain <select> dropdown with a horizontal,
  scrollable avatar-chip picker (avatar + name + role), the active chip tinted
  with the familiar's accent. Scales to a large roster without stealing a left
  rail from the detail pane.
- New per-familiar "Projects" tab (familiar-studio-projects-tab.tsx): the grant
  matrix, access requests, and recent access decisions — all scoped to the
  selected familiar. Grant toggles POST/DELETE /api/project-grants with only the
  target familiar + project; proposals resolve via PATCH with only the decision;
  the supreme (all-access) familiar is surfaced, not toggle-able.
- Remove the standalone Settings → Permissions section (and the now-dead
  settings-permissions.tsx); its search keywords fold into the Familiars entry so
  search still routes there. Shared pure logic stays in permissions-console.ts.
- Studio context gains a "projects" tab; the Workspace drawer degrades it to
  Identity (Settings-only tab, like brain/vault).
- Replace the unused legacy __list/__item/__select CSS with the new picker/chip
  styles.
